% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GNPriv.R, R/GNPrivConfiguration.R
\docType{class}
\name{GNPrivConfiguration}
\alias{GNPriv}
\alias{GNPrivConfiguration}
\title{A GeoNetwork privilege configuration}
\format{\code{\link{R6Class}} object.}
\usage{
GNPriv

GNPrivConfiguration
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoNetwork Privilege configuration

Object of \code{\link{R6Class}} for modelling a GeoNetwork Privilege configuration
}
\description{
This class is an utility to configure privileges

This class is an utility to configure privileges
}
\details{
GeoNetwork REST API - GeoNetwork privilege configuration

GeoNetwork REST API - GeoNetwork privilege configuration
}
\section{Methods}{

\describe{
 \item{\code{new()}}{
   This method is used to instantiate a GNPriv object.
 }
}
}

\section{Methods}{

\describe{
 \item{\code{new()}}{
   This method is used to instantiate a GNPrivConfiguration object.
 }
 \item{\code{setPrivileges(group, privileges)}}{
   Sets the operation privileges for a particular group. Allowed group values
   are "guest","intranet" and "all". Allowed values for operation privileges
   are "view", "download", "editing", "notify", "dynamic" and "featured".
 }
}
}
\examples{
\dontrun{
 priv <- GNPriv$new(
}
\dontrun{
 pcfg <- GNPrivConfiguration$new()
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GeoNetwork}
\keyword{configuration}
\keyword{privilege}

