% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geonapi.R
\docType{package}
\name{geonapi}
\alias{geonapi}
\alias{geonapi-package}
\title{GeoNetwork REST API R Interface}
\description{
Provides an R interface to the GeoNetwork REST API, allowing to upload
and publish metadata in a GeoNetwork web-application and exposteit to OGC Web-Services.
(Catalogue Service for the Web).
}
\details{
\tabular{ll}{
   Package: \tab geonapi\cr
   Type: \tab Package\cr
   Version: \tab 0.1-0\cr
   Date: \tab 2017-05-16\cr
   License: \tab MIT\cr
   LazyLoad: \tab yes\cr
 }
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}

