% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geonapi.R
\docType{package}
\name{geonapi}
\alias{geonapi}
\alias{geonapi-package}
\title{'GeoNetwork' API R Interface}
\description{
Provides an R interface to the 'GeoNetwork' API (<https://geonetwork-opensource.org/#api>)
allowing to upload and publish metadata in a 'GeoNetwork' web-application and expose it to OGC CSW 
Web-Services (Catalogue Service for the Web).
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
