% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geonapi.R
\docType{package}
\name{geonapi}
\alias{geonapi}
\alias{geonapi-package}
\title{'GeoNetwork' API R Interface}
\description{
Provides an R interface to the 'GeoNetwork' API (<https://geonetwork-opensource.org/#api>)
allowing to upload and publish metadata in a 'GeoNetwork' web-application and exposte it to OGC CSW 
Web-Services (Catalogue Service for the Web).
}
\details{
\tabular{ll}{
   Package: \tab geonapi\cr
   Type: \tab Package\cr
   Version: \tab 0.2-2\cr
   Date: \tab 2019-10-18\cr
   License: \tab MIT\cr
   LazyLoad: \tab yes\cr
 }
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
