% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNPriv.R, R/GNPrivConfiguration.R
\docType{class}
\name{GNPrivConfiguration}
\alias{GNPrivConfiguration}
\alias{GNPriv}
\title{A GeoNetwork privilege configuration}
\format{
\code{\link{R6Class}} object.

\code{\link[R6]{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoNetwork Privilege configuration

Object of \code{\link[R6]{R6Class}} for modelling a GeoNetwork Privilege configuration
}
\description{
This class is an utility to configure privileges

This class is an utility to configure privileges
}
\details{
GeoNetwork REST API - GeoNetwork privilege configuration

GeoNetwork REST API - GeoNetwork privilege configuration
}
\examples{
\dontrun{
 priv <- GNPriv$new(group="all", privileges=c("view","dynamic","featured"))
}

\dontrun{
 pcfg <- GNPrivConfiguration$new()
 pcfg$setPrivileges("all", c("view","dynamic","featured"))
}

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GeoNetwork}
\keyword{configuration}
\keyword{privilege}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{group}}{group}

\item{\code{privileges}}{privileges}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GNPriv-new}{\code{GNPriv$new()}}
\item \href{#method-GNPriv-clone}{\code{GNPriv$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GNPriv-new"></a>}}
\if{latex}{\out{\hypertarget{method-GNPriv-new}{}}}
\subsection{Method \code{new()}}{
Initializes a \link{GNPriv} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GNPriv$new(group, privileges)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group}}{group}

\item{\code{privileges}}{privileges}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GNPriv-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GNPriv-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GNPriv$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{privileges}}{privileges}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GNPrivConfiguration-new}{\code{GNPrivConfiguration$new()}}
\item \href{#method-GNPrivConfiguration-setPrivileges}{\code{GNPrivConfiguration$setPrivileges()}}
\item \href{#method-GNPrivConfiguration-clone}{\code{GNPrivConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GNPrivConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-GNPrivConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Initializes an object of class \link{GNPrivConfiguration}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GNPrivConfiguration$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GNPrivConfiguration-setPrivileges"></a>}}
\if{latex}{\out{\hypertarget{method-GNPrivConfiguration-setPrivileges}{}}}
\subsection{Method \code{setPrivileges()}}{
Sets the operation privileges for a particular group. Allowed group values
   are "guest","intranet" and "all". Allowed values for operation privileges
   are "view", "download", "editing", "notify", "dynamic" and "featured".
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GNPrivConfiguration$setPrivileges(group, privileges)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group}}{group}

\item{\code{privileges}}{privileges}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GNPrivConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GNPrivConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GNPrivConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
