% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GNRESTRequest.R
\docType{class}
\name{GNRESTRequest}
\alias{GNRESTRequest}
\title{GeoNetwork REST API REST Request}
\format{\code{\link{R6Class}} object.}
\usage{
GNRESTRequest
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoNetwork REST request
}
\description{
GeoNetwork REST API REST Request
}
\section{Abstract Methods}{

\describe{
 \item{\code{new()}}{
   This method is used to instantiate a GNRESTRequest
 }
 \item{\code{setChild(key,value)}}{
   Sets a child element
 }
 \item{\code{encode()}}{
   Encodes a GNRESTRequest R6 object to XML representation
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{geonetwork}
\keyword{rest}

