% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNVersion.R
\docType{class}
\name{GNVersion}
\alias{GNVersion}
\title{A GeoNetwork version}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoNetwork version
}
\description{
This class is an utility wrap the Geonetwork version
}
\details{
GeoNetwork REST API - GeoNetwork Version
}
\section{Methods}{

\describe{
 \item{\code{new(version)}}{
   This method is used to instantiate a GNVersion object.
 }
 \item{\code{lowerThan(version)}}{
   Compares to a version and returns TRUE if it is lower, FALSE otherwise
 }
 \item{\code{greaterThan(version)}}{
   Compares to a version and returns TRUE if it is greater, FALSE otherwise
 }
 \item{\code{equalTo(version)}}{
   Compares to a version and returns TRUE if it is equal, FALSE otherwise
 }
}
}

\examples{
\dontrun{
version <- GNVersion$new("2.6.4")
}

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GeoNetwork}
\keyword{version}
